Moving out types into respective files
