+++ libs/types/Cargo.toml
@@ -13,6 +13,7 @@ targets = ['x86_64-unknown-linux-gnu']
 [dependencies]
 bitflags = { version = "1.3", default-features = false }
 codec = { package = 'parity-scale-codec', version = '3.0.0', features = ['derive'], default-features = false }
+rev_slice = { version = "0.1.5", default-features = false }
 scale-info = { version = "2.0", default-features = false, features = ["derive"] }
 serde = { version = "1.0.119" }
 
@@ -24,6 +25,7 @@ sp-runtime = { git = "https://github.com/paritytech/substrate", default-features
 sp-std = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.29" }
 
 orml-asset-registry = { git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.29" }
+orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.29" }
 polkadot-parachain = { git = "https://github.com/paritytech/polkadot", default-features = false, branch = "release-v0.9.29" }
 xcm = { git = "https://github.com/paritytech/polkadot", default-features = false, branch = "release-v0.9.29" }
