+++ libs/types/src/adjustments.rs
@@ -10,11 +10,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-//! Mocks of traits for usage in pallet-tests.
-//! Also does contain implementations for Substrate based traits for testing
-
-pub mod accountant;
-pub mod authority_origin;
-pub mod nav;
-pub mod order_manager;
-pub mod orml_asset_registry;
+pub enum Adjustment<Amount> {
+	Increase(Amount),
+	Decrease(Amount),
+}
