+++ libs/types/src/fee_keys.rs
@@ -1,0 +1,54 @@
+// Copyright 2021 Centrifuge Foundation (centrifuge.io).
+//
+// This file is part of the Centrifuge chain project.
+// Centrifuge is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version (see http://www.gnu.org/licenses).
+// Centrifuge is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+use cfg_primitives::types::Balance;
+use cfg_traits::InvestmentProperties;
+use codec::{Decode, Encode, MaxEncodedLen};
+use frame_support::{traits::UnixTime, RuntimeDebug};
+use scale_info::{build::Fields, Path, Type, TypeInfo};
+#[cfg(feature = "std")]
+use serde::{Deserialize, Serialize};
+use sp_runtime::{traits::Zero, Perquintill};
+use sp_std::{
+	cmp::{Ord, PartialEq, PartialOrd},
+	marker::PhantomData,
+};
+
+/// Different fees keys available.
+/// Each variant represents a balance previously determined and configured.
+#[derive(Encode, Decode, Clone, Copy, PartialEq, RuntimeDebug, TypeInfo)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+pub enum FeeKey {
+	/// Key to identify the balance reserved for the author.
+	/// See more at `pallet-anchors`
+	AnchorsCommit,
+
+	/// Key to identify the balance reserved for the deposit.
+	/// See more at `pallet-anchors`
+	AnchorsPreCommit,
+
+	/// Key to identify the balance reserved for burning.
+	/// See more at `pallet-bridge`
+	BridgeNativeTransfer,
+
+	/// Key to identify the balance reserved for burning.
+	/// See more at `pallet-nft`
+	NftProofValidation,
+}
+
+/// Only needed for initializing the runtime benchmark with some value.
+#[cfg(feature = "runtime-benchmarks")]
+impl Default for FeeKey {
+	fn default() -> Self {
+		FeeKey::AnchorsCommit
+	}
+}
