+++ libs/types/src/time.rs
@@ -1,0 +1,49 @@
+// Copyright 2021 Centrifuge Foundation (centrifuge.io).
+//
+// This file is part of the Centrifuge chain project.
+// Centrifuge is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version (see http://www.gnu.org/licenses).
+// Centrifuge is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+use cfg_primitives::types::Balance;
+use cfg_traits::InvestmentProperties;
+use codec::{Decode, Encode, MaxEncodedLen};
+use frame_support::{traits::UnixTime, RuntimeDebug};
+use scale_info::{build::Fields, Path, Type, TypeInfo};
+#[cfg(feature = "std")]
+use serde::{Deserialize, Serialize};
+use sp_runtime::{traits::Zero, Perquintill};
+use sp_std::{
+	cmp::{Ord, PartialEq, PartialOrd},
+	marker::PhantomData,
+};
+
+/// A struct we need as the pallets implementing trait Time
+/// do not implement TypeInfo. This wraps this and implements everything manually.
+#[derive(Encode, Decode, Eq, PartialEq, Debug, Clone)]
+pub struct TimeProvider<T>(PhantomData<T>);
+
+impl<T> UnixTime for TimeProvider<T>
+where
+	T: UnixTime,
+{
+	fn now() -> core::time::Duration {
+		<T as UnixTime>::now()
+	}
+}
+
+impl<T> TypeInfo for TimeProvider<T> {
+	type Identity = ();
+
+	fn type_info() -> Type {
+		Type::builder()
+			.path(Path::new("TimeProvider", module_path!()))
+			.docs(&["A wrapper around a T that provides a trait Time implementation. Should be filtered out."])
+			.composite(Fields::unit())
+	}
+}
