+++ libs/types/src/tokens.rs
@@ -57,21 +57,6 @@ impl From<u32> for CurrencyId {
 	}
 }
 
-/// A Currency that is solely used by tranches.
-///
-/// We distinguish here between the enum variant CurrencyId::Tranche(PoolId, TranchId)
-/// in order to be able to have a clear separation of concerns. This enables us
-/// to use the `TrancheCurrency` type separately where solely this enum variant would be
-/// relevant. Most notably, in the `struct Tranche`.
-#[derive(
-	Clone, Copy, PartialOrd, Ord, PartialEq, Eq, Debug, Encode, Decode, TypeInfo, MaxEncodedLen,
-)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-pub struct TrancheCurrency {
-	pool_id: PoolId,
-	tranche_id: TrancheId,
-}
-
 impl From<TrancheCurrency> for CurrencyId {
 	fn from(x: TrancheCurrency) -> Self {
 		CurrencyId::Tranche(x.pool_id, x.tranche_id)
