+++ pallets/pool-system/src/lib.rs
@@ -59,126 +59,44 @@ mod mock;
 mod solution;
 #[cfg(test)]
 mod tests;
-mod tranche;
 pub mod weights;
 
-/// A convenience struct to easily pass around the accumulated orders
-/// for all tranches, which is of sole interest to the pool.
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-struct SummarizedOrders<Balance> {
-	// The accumulated order amounts of all investments
-	acc_invest_orders: Balance,
-	// The accumulated order amounts of all redemptions
-	//
-	// NOTE: Already denominated in the pool_currency!
-	acc_redeem_orders: Balance,
-	// Invest orders per tranche
-	//
-	// NOTE: Sorted from residual-to-non-residual
-	invest_orders: Vec<Balance>,
-	// Redeem orders per tranche
-	//
-	// NOTE: Sorted from residual-to-non-residual
-	redeem_orders: Vec<Balance>,
-}
+/// Types alias for EpochExecutionTranche
+#[allow(dead_code)]
+pub type EpochExecutionTrancheOf<T> = EpochExecutionTranche<
+	<T as Config>::Balance,
+	<T as Config>::Rate,
+	<T as Config>::TrancheWeight,
+	<T as Config>::TrancheCurrency,
+>;
 
-impl<Balance: Zero + PartialEq + Eq + Copy> SummarizedOrders<Balance> {
-	fn all_are_zero(&self) -> bool {
-		self.acc_invest_orders == Zero::zero() && self.acc_redeem_orders == Zero::zero()
-	}
+#[allow(dead_code)]
+/// Type alias for EpochExecutionTranches
+pub type EpochExecutionTranchesOf<T> = EpochExecutionTranches<
+	<T as Config>::Balance,
+	<T as Config>::Rate,
+	<T as Config>::TrancheWeight,
+	<T as Config>::TrancheCurrency,
+>;
 
-	fn invest_redeem_residual_top(&self) -> Vec<(Balance, Balance)> {
-		self.invest_orders
-			.iter()
-			.zip(&self.redeem_orders)
-			.map(|(invest, redeem)| (*invest, *redeem))
-			.collect::<Vec<_>>()
-	}
-}
+/// Types alias for Tranches
+pub type TranchesOf<T> = Tranches<
+	<T as Config>::Balance,
+	<T as Config>::Rate,
+	<T as Config>::TrancheWeight,
+	<T as Config>::TrancheCurrency,
+	<T as Config>::TrancheId,
+	<T as Config>::PoolId,
+>;
 
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct PoolDetails<
-	CurrencyId,
-	TrancheCurrency,
-	EpochId,
-	Balance,
-	Rate,
-	MetaSize,
-	Weight,
-	TrancheId,
-	PoolId,
-> where
-	MetaSize: Get<u32> + Copy,
-	Rate: FixedPointNumber<Inner = Balance>,
-	Balance: FixedPointOperand,
-{
-	/// Currency that the pool is denominated in (immutable).
-	pub currency: CurrencyId,
-	/// List of tranches, ordered junior to senior.
-	pub tranches: Tranches<Balance, Rate, Weight, TrancheCurrency, TrancheId, PoolId>,
-	/// Details about the parameters of the pool.
-	pub parameters: PoolParameters,
-	/// Metadata that specifies the pool.
-	pub metadata: Option<BoundedVec<u8, MetaSize>>,
-	/// The status the pool is currently in.
-	pub status: PoolStatus,
-	/// Details about the epochs of the pool.
-	pub epoch: EpochState<EpochId>,
-	/// Details about the reserve (unused capital) in the pool.
-	pub reserve: ReserveDetails<Balance>,
-}
-
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub enum PoolStatus {
-	Open,
-}
-
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct ReserveDetails<Balance> {
-	/// Investments will be allowed up to this amount.
-	pub max: Balance,
-	/// Current total amount of currency in the pool reserve.
-	pub total: Balance,
-	/// Current reserve that is available for originations.
-	pub available: Balance,
-}
-
-impl<Balance> ReserveDetails<Balance>
-where
-	Balance: AtLeast32BitUnsigned + Copy + From<u64>,
-{
-	fn deposit_from_epoch<BalanceRatio, Weight, TrancheCurrency>(
-		&mut self,
-		epoch_tranches: &EpochExecutionTranches<Balance, BalanceRatio, Weight, TrancheCurrency>,
-		solution: &[TrancheSolution],
-	) -> DispatchResult
-	where
-		Weight: Copy + From<u128>,
-		BalanceRatio: Copy,
-	{
-		let executed_amounts = epoch_tranches.fulfillment_cash_flows(solution)?;
-
-		// Update the total/available reserve for the new total value of the pool
-		let mut acc_investments = Balance::zero();
-		let mut acc_redemptions = Balance::zero();
-		for (invest, redeem) in executed_amounts.iter() {
-			acc_investments = acc_investments
-				.checked_add(invest)
-				.ok_or(ArithmeticError::Overflow)?;
-			acc_redemptions = acc_redemptions
-				.checked_add(redeem)
-				.ok_or(ArithmeticError::Overflow)?;
-		}
-		self.total = self
-			.total
-			.checked_add(&acc_investments)
-			.ok_or(ArithmeticError::Overflow)?
-			.checked_sub(&acc_redemptions)
-			.ok_or(ArithmeticError::Underflow)?;
-
-		Ok(())
-	}
-}
+#[allow(dead_code)]
+/// Types alias for Tranche
+pub type TrancheOf<T> = Tranche<
+	<T as Config>::Balance,
+	<T as Config>::Rate,
+	<T as Config>::TrancheWeight,
+	<T as Config>::TrancheCurrency,
+>;
 
 #[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
 pub struct EpochState<EpochId> {
@@ -188,28 +106,6 @@ pub struct EpochState<EpochId> {
 	pub last_closed: Moment,
 	/// Last epoch that was executed.
 	pub last_executed: EpochId,
-}
-
-#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct PoolParameters {
-	/// Minimum duration for an epoch.
-	pub min_epoch_time: Moment,
-	/// Maximum time between the NAV update and the epoch closing.
-	pub max_nav_age: Moment,
-}
-
-#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct PoolChanges<Rate, MaxTokenNameLength, MaxTokenSymbolLength, MaxTranches>
-where
-	MaxTokenNameLength: Get<u32>,
-	MaxTokenSymbolLength: Get<u32>,
-	MaxTranches: Get<u32>,
-{
-	pub tranches: Change<BoundedVec<TrancheUpdate<Rate>, MaxTranches>>,
-	pub tranche_metadata:
-		Change<BoundedVec<TrancheMetadata<MaxTokenNameLength, MaxTokenSymbolLength>, MaxTranches>>,
-	pub min_epoch_time: Change<Moment>,
-	pub max_nav_age: Change<Moment>,
 }
 
 #[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
@@ -223,95 +119,6 @@ where
 	pub scheduled_time: Moment,
 }
 
-impl<CurrencyId, TrancheCurrency, EpochId, Balance, Rate, MetaSize, Weight, TrancheId, PoolId>
-	PoolDetails<
-		CurrencyId,
-		TrancheCurrency,
-		EpochId,
-		Balance,
-		Rate,
-		MetaSize,
-		Weight,
-		TrancheId,
-		PoolId,
-	> where
-	Balance: FixedPointOperand + BaseArithmetic + Unsigned + From<u64>,
-	CurrencyId: Copy,
-	EpochId: BaseArithmetic,
-	MetaSize: Get<u32> + Copy,
-	PoolId: Copy + Encode,
-	Rate: FixedPointNumber<Inner = Balance>,
-	TrancheCurrency: Copy + cfg_traits::TrancheCurrency<PoolId, TrancheId>,
-	TrancheId: Clone + From<[u8; 16]> + PartialEq,
-	Weight: Copy + From<u128>,
-{
-	pub fn start_next_epoch(&mut self, now: Moment) -> DispatchResult {
-		self.epoch.current += One::one();
-		self.epoch.last_closed = now;
-		// TODO: Remove and set state rather to EpochClosing or similar
-		// Set available reserve to 0 to disable originations while the epoch is closed but not executed
-		self.reserve.available = Zero::zero();
-
-		Ok(())
-	}
-
-	fn execute_previous_epoch(&mut self) -> DispatchResult {
-		self.reserve.available = self.reserve.total;
-		self.epoch.last_executed += One::one();
-		Ok(())
-	}
-
-	pub fn essence<
-		T: Config<
-			CurrencyId = CurrencyId,
-			Balance = Balance,
-			TrancheCurrency = TrancheCurrency,
-			Rate = Rate,
-		>,
-	>(
-		&self,
-	) -> Result<PoolEssenceOf<T>, DispatchError> {
-		let mut tranches: Vec<
-			TrancheEssence<
-				T::TrancheCurrency,
-				T::Rate,
-				T::MaxTokenNameLength,
-				T::MaxTokenSymbolLength,
-			>,
-		> = Vec::new();
-
-		for tranche in self.tranches.residual_top_slice().iter() {
-			let metadata = T::AssetRegistry::metadata(&self.currency).ok_or(AssetMetadata {
-				decimals: 0,
-				name: Vec::new(),
-				symbol: Vec::new(),
-				existential_deposit: (),
-				location: None,
-				additional: (),
-			});
-
-			tranches.push(TrancheEssence {
-				currency: tranche.currency.into(),
-				tranche_type: tranche.tranche_type.into(),
-				metadata: TrancheMetadata {
-					token_name: BoundedVec::try_from(metadata.clone().unwrap().name)
-						.unwrap_or(BoundedVec::default()),
-					token_symbol: BoundedVec::try_from(metadata.unwrap().symbol)
-						.unwrap_or(BoundedVec::default()),
-				},
-			});
-		}
-
-		Ok(PoolEssence {
-			currency: self.currency,
-			max_reserve: self.reserve.max.into(),
-			max_nav_age: self.parameters.max_nav_age,
-			min_epoch_time: self.parameters.min_epoch_time,
-			tranches,
-		})
-	}
-}
-
 /// The information for a currently executing epoch
 #[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
 pub struct EpochExecutionInfo<Balance, BalanceRatio, EpochId, Weight, BlockNumber, TrancheCurrency>
@@ -323,55 +130,6 @@ pub struct EpochExecutionInfo<Balance, BalanceRatio, EpochId, Weight, BlockNumbe
 	tranches: EpochExecutionTranches<Balance, BalanceRatio, Weight, TrancheCurrency>,
 	best_submission: Option<EpochSolution<Balance>>,
 	challenge_period_end: Option<BlockNumber>,
-}
-
-/// Information about the deposit that has been taken to create a pool
-#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, Default, TypeInfo)]
-pub struct PoolDepositInfo<AccountId, Balance> {
-	pub depositor: AccountId,
-	pub deposit: Balance,
-}
-
-/// The core metadata about the pool which we can attach to an event
-#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct PoolEssence<
-	CurrencyId,
-	Balance,
-	TrancheCurrency,
-	Rate,
-	MaxTokenNameLength,
-	MaxTokenSymbolLength,
-> where
-	CurrencyId: Copy,
-	MaxTokenNameLength: Get<u32>,
-	MaxTokenSymbolLength: Get<u32>,
-{
-	/// Currency that the pool is denominated in (immutable).
-	pub currency: CurrencyId,
-	/// The maximum allowed reserve on a given pool
-	pub max_reserve: Balance,
-	/// Maximum time between the NAV update and the epoch closing.
-	pub max_nav_age: Moment,
-	/// Minimum duration for an epoch.
-	pub min_epoch_time: Moment,
-	/// Tranches on a pool
-	pub tranches:
-		Vec<TrancheEssence<TrancheCurrency, Rate, MaxTokenNameLength, MaxTokenSymbolLength>>,
-}
-
-/// The core metadata about a tranche which we can attach to an event
-#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct TrancheEssence<TrancheCurrency, Rate, MaxTokenNameLength, MaxTokenSymbolLength>
-where
-	MaxTokenNameLength: Get<u32>,
-	MaxTokenSymbolLength: Get<u32>,
-{
-	/// Currency that the tranche is denominated in
-	pub currency: TrancheCurrency,
-	/// Type of the tranche (Residual or NonResidual)
-	pub tranche_type: TrancheType<Rate>,
-	/// Metadata of a Tranche
-	pub metadata: TrancheMetadata<MaxTokenNameLength, MaxTokenSymbolLength>,
 }
 
 /// Type alias to ease function signatures
